fix(Android): Rename Android test app package to com.microsoft.reacttestapp
