+++ android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.react.testapp">
+    package="com.microsoft.reacttestapp">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -18,14 +18,14 @@
 
         <activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
 
-        <activity android:name="com.react.testapp.MainActivity">
+        <activity android:name="com.microsoft.reacttestapp.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name="com.react.testapp.component.ComponentActivity" />
+        <activity android:name="com.microsoft.reacttestapp.component.ComponentActivity" />
     </application>
 
 </manifest>
