+++ android/app/src/main/java/com/microsoft/reacttestapp/manifest/Manifest.kt
@@ -5,12 +5,21 @@
 // LICENSE file in the root directory of this source tree.
 //
 
-package com.react.testapp.component
+package com.microsoft.reacttestapp.manifest
 
 import android.os.Bundle
+import com.squareup.moshi.JsonClass
 
-data class ComponentViewModel(
+@JsonClass(generateAdapter = true)
+data class Manifest(
     val name: String,
     val displayName: String,
+    val components: List<Component>
+)
+
+@JsonClass(generateAdapter = true)
+data class Component(
+    val appKey: String,
+    val displayName: String?,
     val initialProperties: Bundle?
 )
