+++ android/app/src/main/java/com/microsoft/reacttestapp/react/ReactBundleNameProvider.kt
@@ -1,4 +1,4 @@
-package com.react.testapp.react
+package com.microsoft.reacttestapp.react
 
 import android.content.Context
 import javax.inject.Inject
